Move MessagePlugin to event controller_front_send_response_before
